requirements.txt
# Flask and Extensions
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
flask-restx==1.3.0

# MongoDB
pymongo==4.6.1
pymongo[srv]==4.6.1

# Password Hashing
bcrypt==4.1.2

# Azure Blob Storage
azure-storage-blob==12.19.0

# Email
secure-smtplib==0.1.1

# PDF Generation
reportlab==4.0.8
PyPDF2==3.0.1

# Environment Variables
python-dotenv==1.0.0

# Date/Time
python-dateutil==2.8.2

# Validation
marshmallow==3.20.1
email-validator==2.1.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0

# Utilities
Werkzeug==3.0.1

# Payment Gateway
razorpay==1.4.1

# Scheduler
schedule==1.2.1

# HTTP Requests (for Azure OpenAI)
requests==2.31.0
